chore: fix linter
